Problem: lack Elixir, way to run benchmark once for debug, Haskell LLVM support
Solution: we add simplistic script called `runOnce.sh` that just runs the
benchmark 1000 times under `time`. Also, this commit adds an Elixir module that
is ran using `elixir` binary. Futher work is required to make it into a
benchmark for stripped Escript which should reduce the startup time, but this
is a good enough first approximation for user-facing programs written in
Elixir.
Finally, in this commit we make sure that the Haskell module is called `Main`
to be able to compile it properly with `-o` key, for instance while doing `gcc
-fllvm` compilation.